# docs/swagger.yaml
basePath: /v1
host: localhost:8080
info:
  contact: {}
  description: This is a sample API
  title: My API
  version: "1.0"
paths:
  /:app:
    delete:
      consumes:
      - application/json
      description: Pagination DeleteApp
      operationId: ApiV1LocalizationDeleteAppPagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination DeleteApp
      tags:
      - app/
    get:
      consumes:
      - application/json
      description: Pagination ReadApp
      operationId: ApiV1LocalizationReadAppPagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: app
        in: path
        name: app
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination ReadApp
      tags:
      - app/
    post:
      consumes:
      - application/json
      description: Pagination CreateApp
      operationId: ApiV1LocalizationCreateAppPagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination CreateApp
      tags:
      - app/
    put:
      consumes:
      - application/json
      description: Pagination UpdateApp
      operationId: ApiV1LocalizationUpdateAppPagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination UpdateApp
      tags:
      - app/
  /:app/:module:
    delete:
      consumes:
      - application/json
      description: Pagination DeleteModule
      operationId: ApiV1LocalizationDeleteModulePagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: module
        in: path
        name: module
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination DeleteModule
      tags:
      - app/module
    get:
      consumes:
      - application/json
      description: Pagination ReadModule
      operationId: ApiV1LocalizationReadModulePagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: module
        in: path
        name: module
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination ReadModule
      tags:
      - app/module
    post:
      consumes:
      - application/json
      description: Pagination CreateModule
      operationId: ApiV1LocalizationCreateModulePagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: module
        in: path
        name: module
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination CreateModule
      tags:
      - app/module
    put:
      consumes:
      - application/json
      description: Pagination UpdateModule
      operationId: ApiV1LocalizationUpdateModulePagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: module
        in: path
        name: module
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination UpdateModule
      tags:
      - app/module
  /:app/:module/:language:
    delete:
      consumes:
      - application/json
      description: Pagination DeleteLanguage
      operationId: ApiV1LocalizationDeleteLanguagePagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: module
        in: path
        name: module
        required: true
        type: string
      - description: language
        in: path
        name: language
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination DeleteLanguage
      tags:
      - app/module/language
    get:
      consumes:
      - application/json
      description: Pagination ReadLanguage
      operationId: ApiV1LocalizationReadLanguagePagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: module
        in: path
        name: module
        required: true
        type: string
      - description: language
        in: path
        name: language
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination ReadLanguage
      tags:
      - app/module/language
    post:
      consumes:
      - application/json
      description: Pagination CreateLanguage
      operationId: ApiV1LocalizationCreateLanguagePagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: module
        in: path
        name: module
        required: true
        type: string
      - description: language
        in: path
        name: language
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination CreateLanguage
      tags:
      - app/module/language
    put:
      consumes:
      - application/json
      description: Pagination UpdateLanguage
      operationId: ApiV1LocalizationUpdateLanguagePagination
      parameters:
      - description: app
        in: path
        name: app
        required: true
        type: string
      - description: module
        in: path
        name: module
        required: true
        type: string
      - description: language
        in: path
        name: language
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Pagination UpdateLanguage
      tags:
      - app/module/language
swagger: "2.0"